Extract home client content into dedicated component
